--- src/components/Layout/MainSidebar/index.ts
@@ -1,1 +1,0 @@
-export { MainSidebar } from './MainSidebar';
