+++ src/components/UI/Loader/Loader.module.scss
@@ -14,5 +14,5 @@
     height: 100%;
     filter: blur(12px);
     background: rgba(121, 156, 212, 0.5);
-    z-index: 5;
+    z-index: 100;
 }
