--- src/pages/auth-page/index.ts
@@ -1,1 +1,0 @@
-export { AuthPage } from './auth-page';
