--- src/pages/feedback-page/index.ts
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export const FeedbackPage = React.lazy(() => import('./feedback-page'));
-
-
