+++ src/pages/index.ts
@@ -1,3 +1,8 @@
-export { MainPage } from './main-page';
-export { AuthPage } from './auth-page';
-export { FeedbackPage } from './feedback-page'
+import { lazy } from 'react';
+
+
+
+export { AuthPage } from './auth-page/auth-page';
+
+export const MainPage = lazy(() => import('./main-page/main-page'));
+export const FeedbackPage = lazy(() => import('./feedback-page/feedback-page'));
