+++ src/pages/main-page/MainContent/MainContent.tsx
