+++ src/pages/main-page/MainContent/index.ts
