+++ src/pages/main-page/MainFooter/MainFooter.tsx
