+++ src/pages/main-page/MainFooter/index.ts
