+++ src/pages/main-page/MainHeader/MainHeader.tsx
