+++ src/pages/main-page/MainHeader/index.ts
