+++ tsconfig.json
@@ -61,6 +61,9 @@
             "@assets/*": [
                 "./src/assets/*"
             ],
+            "@helpers/*": [
+                "./src/helpers/*"
+            ],
         },
         "plugins": [
             {
