+++ vite.config.ts
@@ -24,6 +24,7 @@ export default defineConfig({
             '@services': path.resolve(__dirname, 'src/services'),
             '@assets': path.resolve(__dirname, 'src/assets'),
             '@interfaces': path.resolve(__dirname, 'src/interfaces'),
+            '@helpers': path.resolve(__dirname, 'src/helpers'),
         },
     },
 });
